edit documents function, internet stability check,and dialogs

+++ app/src/main/java/com/example/company_app/adapters/workDayAdapter.kt
@@ -1,12 +1,15 @@
 package com.example.company_app.adapters
 
 import android.app.AlertDialog
+import android.content.Context
+import android.net.ConnectivityManager
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
+import android.widget.Toast
 import androidx.recyclerview.widget.RecyclerView
 import com.example.company_app.R
 import com.example.company_app.classes.objectData
@@ -15,9 +18,14 @@ class workDayAdapter(private val manifestoList: ArrayList<objectData>) :
     RecyclerView.Adapter<workDayAdapter.workDayViewHolder>() {
 
     private var onDeleteClickListener: OnDeleteClickListener? = null
+    private var onEditClickListener: OnEditClickListener? = null
 
     interface OnDeleteClickListener {
         fun onDeleteClick(manifesto: objectData)
+    }
+
+    interface OnEditClickListener {
+        fun onEditClick(manifesto: objectData)
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): workDayViewHolder {
@@ -38,8 +46,13 @@ class workDayAdapter(private val manifestoList: ArrayList<objectData>) :
         onDeleteClickListener = listener
     }
 
+    fun setOnEditClickListener(listener: OnEditClickListener) {
+        onEditClickListener = listener
+    }
+
     inner class workDayViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         private val bin: ImageView = itemView.findViewById(R.id.binImg)
+        private val editImg: ImageView = itemView.findViewById(R.id.editImg)
         private val name: TextView = itemView.findViewById(R.id.tv_Name)
         private val hours: TextView = itemView.findViewById(R.id.tvHours)
         private val comment: TextView = itemView.findViewById(R.id.tvComment)
@@ -54,21 +67,60 @@ class workDayAdapter(private val manifestoList: ArrayList<objectData>) :
             bin.setOnClickListener {
                 showDeleteConfirmationDialog(manifesto)
             }
+            editImg.setOnClickListener {
+                showEditConfirmationDialog(manifesto)
+            }
         }
 
         private fun showDeleteConfirmationDialog(manifesto: objectData) {
             val context = itemView.context
-            val alertDialogBuilder = AlertDialog.Builder(context)
-            alertDialogBuilder.setTitle("Delete Confirmation")
-            alertDialogBuilder.setMessage("Are you sure you want to delete this item?")
-            alertDialogBuilder.setPositiveButton("Yes") { _, _ ->
-                onDeleteClickListener?.onDeleteClick(manifesto)
+
+            if (isNetworkConnected()) {
+                val alertDialogBuilder = AlertDialog.Builder(context)
+                alertDialogBuilder.setTitle("Delete Confirmation")
+                alertDialogBuilder.setMessage("Are you sure you want to delete this item?")
+                alertDialogBuilder.setPositiveButton("Yes") { _, _ ->
+                    onDeleteClickListener?.onDeleteClick(manifesto)
+                }
+                alertDialogBuilder.setNegativeButton("No") { dialog, _ ->
+                    dialog.dismiss()
+                }
+                val alertDialog = alertDialogBuilder.create()
+                alertDialog.show()
+            } else {
+                Toast.makeText(context, "No stable internet connection", Toast.LENGTH_SHORT).show()
             }
-            alertDialogBuilder.setNegativeButton("No") { dialog, _ ->
-                dialog.dismiss()
-            }
-            val alertDialog = alertDialogBuilder.create()
-            alertDialog.show()
         }
+
+        private fun showEditConfirmationDialog(manifesto: objectData) {
+            val context = itemView.context
+
+            if (isNetworkConnected()) {
+                val alertDialogBuilder = AlertDialog.Builder(context)
+                alertDialogBuilder.setTitle("Edit Confirmation")
+                alertDialogBuilder.setMessage("Are you sure you want to edit this item?")
+                alertDialogBuilder.setPositiveButton("Yes") { _, _ ->
+                    onEditClickListener?.onEditClick(manifesto)
+                }
+                alertDialogBuilder.setNegativeButton("No") { dialog, _ ->
+                    dialog.dismiss()
+                }
+                val alertDialog = alertDialogBuilder.create()
+                alertDialog.show()
+            } else {
+                Toast.makeText(context, "No stable internet connection", Toast.LENGTH_SHORT).show()
+            }
+        }
+
+        private fun isNetworkConnected(): Boolean {
+            val connectivityManager = itemView.context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+            val networkInfo = connectivityManager.activeNetworkInfo
+            return networkInfo != null && networkInfo.isConnected
+        }
+
+
+
+
+
     }
 }
